% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckNull_function.R
\name{CheckNulls}
\alias{CheckNulls}
\title{Check Nulls}
\usage{
CheckNulls(x)
}
\arguments{
\item{x}{data}
}
\value{
NA if null, data otherwise
}
\description{
This function checks if data is NULL
}
