% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_global_igraph2tnet_function.R
\name{clustering_global_igraph2tnet}
\alias{clustering_global_igraph2tnet}
\title{Weighted clustering global igraph2tnet}
\usage{
clustering_global_igraph2tnet(gs, alpha)
}
\arguments{
\item{gs}{igraph object}

\item{alpha}{alpha value}
}
\value{
Dataframe of weighted global clustering results results
}
\description{
This function calculates tnet weighted global clustering directly on an igraph object and return a dataframe
}
