% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company_ExtractDirectorsData_function.R
\name{company_ExtractDirectorsData}
\alias{company_ExtractDirectorsData}
\title{Company Directors}
\usage{
company_ExtractDirectorsData(coyno, mkey)
}
\arguments{
\item{coyno}{Company Number}

\item{mkey}{Authorisation key}
}
\value{
Dataframe director information
}
\description{
This function gives director information for a company. Including director name, occupation, date appointed, date resigned and residence.
}
