% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyCentrality_function.R
\name{CompanyCentrality}
\alias{CompanyCentrality}
\title{One-Mode Company Network Centrality Table}
\usage{
CompanyCentrality(gs)
}
\arguments{
\item{gs}{Company Network (igraph object)}
}
\value{
One-Mode Company Network Centrality Table Data frame
}
\description{
This function calculates a number of centrality metrics for the company to company network.
}
