% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyDataExtract_function.R
\name{CompanyDataExtract}
\alias{CompanyDataExtract}
\title{Company Data Extraction}
\usage{
CompanyDataExtract(company_number, mkey)
}
\arguments{
\item{company_number}{Company number}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode, SIC code for the company number
}
\description{
This function extracts company information based on a company number
}
