% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyNetworkProperties_function.R
\name{CompanyNetworkProperties}
\alias{CompanyNetworkProperties}
\title{One-Mode Company Network Properties Table}
\usage{
CompanyNetworkProperties(gs)
}
\arguments{
\item{gs}{Company Network (igraph object)}
}
\value{
One-Mode Company Network Properties Table Data frame
}
\description{
This function calculates network properties for the company to company network.
}
