% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanySearch_function.R
\name{CompanySearch}
\alias{CompanySearch}
\title{Company Search}
\usage{
CompanySearch(company, mkey)
}
\arguments{
\item{company}{Company name search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of all companies matching the search term
}
\description{
This function gives a list of companies, their company numbers and other information that match the company search term
}
