% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanySearch_limit_function.R
\name{CompanySearch_limit}
\alias{CompanySearch_limit}
\title{Company Search Limit}
\usage{
CompanySearch_limit(company, mkey)
}
\arguments{
\item{company}{Company name search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of limited number companies matching the search term (according to best match)
}
\description{
This function gives a list of companies, their company numbers and other information that match the company search term - but only for the first page
}
