% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanySearch_limit_first_function.R
\name{CompanySearch_limit_first}
\alias{CompanySearch_limit_first}
\title{Company Search Limit First}
\usage{
CompanySearch_limit_first(company, mkey)
}
\arguments{
\item{company}{Company name search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of all companies matching the search term
}
\description{
This function gives a list of companies, their company numbers and other information that match the company search term - but only for the first page
}
