% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanySIC_function.R
\name{CompanySIC}
\alias{CompanySIC}
\title{Company SIC}
\usage{
CompanySIC(coyno, mkey)
}
\arguments{
\item{coyno}{Company Number}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name and SIC code
}
\description{
This function gives the company sector description - the SIC code.
}
