% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectorCentrality_function.R
\name{DirectorCentrality}
\alias{DirectorCentrality}
\title{One-Mode Directors Network Centrality Table}
\usage{
DirectorCentrality(gs)
}
\arguments{
\item{gs}{Director Network (igraph object)}
}
\value{
One-Mode Directors Network Centrality Table Data frame
}
\description{
This function calculates a number of centrality metrics for the director to director network.
}
