% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectorNetworkProperties_function.R
\name{DirectorNetworkProperties}
\alias{DirectorNetworkProperties}
\title{One-Mode Directors Network Properties Table}
\usage{
DirectorNetworkProperties(gs)
}
\arguments{
\item{gs}{Director Network (igraph object)}
}
\value{
One-Mode Directors Network Properties Table Data frame
}
\description{
This function calculates network properties for the director to director network.
}
