% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectorSearch_function.R
\name{DirectorSearch}
\alias{DirectorSearch}
\title{Director Search}
\usage{
DirectorSearch(director_name, mkey)
}
\arguments{
\item{director_name}{Director name search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of all companies matching the search term
}
\description{
This function gives a list of direcotrs that match the search term
}
