% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectorSearch_limit_function.R
\name{DirectorSearch_limit}
\alias{DirectorSearch_limit}
\title{Director Search Limit}
\usage{
DirectorSearch_limit(director_name, mkey)
}
\arguments{
\item{director_name}{Director name search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of all companies matching the search term
}
\description{
This function gives a list of direcotrs that match the search term - but only for the first page
}
