% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectorSearch_limit_first_function.R
\name{DirectorSearch_limit_first}
\alias{DirectorSearch_limit_first}
\title{Director Search Limit First}
\usage{
DirectorSearch_limit_first(director_name, mkey)
}
\arguments{
\item{director_name}{Director name search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of all companies matching the search term
}
\description{
This function gives a list of companies, their company numbers and other information that match the company search term - but only for the first item
}
