% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_getlink_function.R
\name{doc_getlink}
\alias{doc_getlink}
\title{Generate Document Download Link}
\usage{
doc_getlink(doc_id, mkey, doc_type)
}
\arguments{
\item{doc_id}{Document number}

\item{mkey}{Authorisation key}

\item{doc_type}{The filetype of the document. Should be one of \code{'application/pdf'},
\code{'application/json'}, \code{'application/xml'}, \code{'application/xhtml+xml'} or \code{'text/csv'}}
}
\value{
Character vector of document location
}
\description{
This function generates the download link for a particular document. Use with \code{doc_meta_extract}
for best results
}
