% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_link_extract_function.R
\name{doc_link_extract}
\alias{doc_link_extract}
\title{Document List Extraction}
\usage{
doc_link_extract(company_number, mkey)
}
\arguments{
\item{company_number}{Company number}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company number, date, docuemnt type, link, meta link  and meta ID
}
\description{
This function provides a list of document numbers & links a company number
}
