% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_meta_extract_function.R
\name{doc_meta_extract}
\alias{doc_meta_extract}
\title{Document Metadata Extraction}
\usage{
doc_meta_extract(doc_id, mkey)
}
\arguments{
\item{doc_id}{Document number}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company number, document number, date, date type,
document type, document format, and created/updated dates
}
\description{
This function returns the metadata of a particular document number
}
