% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty2na_function.R
\name{empty2na}
\alias{empty2na}
\title{empty2na}
\usage{
empty2na(x)
}
\arguments{
\item{x}{data}
}
\value{
NA if empty, data otherwise
}
\description{
Check is an object is empty and return "NA if it is.
}
