% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_ExtractDirectorsData_function.R
\name{indiv_ExtractDirectorsData}
\alias{indiv_ExtractDirectorsData}
\title{Individual Directors}
\usage{
indiv_ExtractDirectorsData(director_id, mkey)
}
\arguments{
\item{director_id}{director id}

\item{mkey}{Authorisation key}
}
\value{
Dataframe director information
}
\description{
This function gives director information for an individuals.Includes what boards they sit on etc.
}
