% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterlockCentrality_function.R
\name{InterlockCentrality}
\alias{InterlockCentrality}
\title{Two-Mode Interlocking Directorates Network Centrality Table}
\usage{
InterlockCentrality(gs)
}
\arguments{
\item{gs}{Interlocking Directorates Network (igraph object)}
}
\value{
Two-Mode Interlocking Directorates Network Centrality Table Data frame
}
\description{
This function calculates degree centrality for the interlocking directorates network.
}
