% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterlockNetworkPLOT_function.R
\name{InterlockNetworkPLOT}
\alias{InterlockNetworkPLOT}
\title{Interlocking Directorates Network}
\usage{
InterlockNetworkPLOT(coynoLIST, mkey, LABEL, NodeSize)
}
\arguments{
\item{coynoLIST}{list of company numbers}

\item{mkey}{Authorisation key}

\item{LABEL}{Node Label - TRUE/FALSE}

\item{NodeSize}{Node Size - default is 6, place number or CENTRALITY (degree centrality)}

\item{YEAR}{Year - put CURRENT for current realisation of the network}
}
\value{
Two-Mode Interlocking Directorates Network - igraph object
}
\description{
This function creates an interlocking directorates network from a list of company number
}
