% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty_function.R
\name{is.empty}
\alias{is.empty}
\title{is.empty}
\usage{
is.empty(x, mode = NULL, ...)
}
\arguments{
\item{x}{data}

\item{mode}{mode}
}
\value{
Answer
}
\description{
This function checks if object is empty
}
