% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeInterlock_function.R
\name{make_interlock}
\alias{make_interlock}
\title{Make an interlocking directorates network}
\usage{
make_interlock(DF, start, end)
}
\arguments{
\item{DF}{Dataframe in the companies house format}

\item{start}{Start Year}

\item{end}{End Year}
}
\value{
Two-Mode Interlocking Directorates Network - igraph object
}
\description{
This function creates an interlocking directorates network from a dataframe with company and director data
}
