% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyCentrality_function.R
\name{one_mode_centrality}
\alias{one_mode_centrality}
\title{One-Mode Network Centrality Table}
\usage{
one_mode_centrality(gs)
}
\arguments{
\item{gs}{Company Network  or Director Network (igraph object)}
}
\value{
One-Mode Company Network Centrality Table Data frame
}
\description{
This function calculates a number of centrality metrics for the one-mode network ie.e either the company or director networks.
}
