% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonSearch_limit_first_function.R
\name{PersonSearch_limit_first}
\alias{PersonSearch_limit_first}
\title{Person Search Limit First}
\usage{
PersonSearch_limit_first(person, mkey)
}
\arguments{
\item{person}{person search term}

\item{mkey}{Authorisation key}
}
\value{
Dataframe listing company name, company number, postcode of limited number companies matching the search term (according to best match)
}
\description{
This function gives a list of companies, their company numbers and other information that match the company search term - but only for the first item
}
