% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateLimitWrap_function.R
\name{RateLimitWrap}
\alias{RateLimitWrap}
\title{Rate Limiting Wrapper}
\usage{
RateLimitWrap(func, ..., buffer = 5, silent = FALSE)
}
\arguments{
\item{func}{The function you want to add rate limiting to}

\item{buffer}{Optional. The amount of requests left before the wrapper will pause. Default is 5}

\item{silent}{Optional. Suppresses printed messages}
}
\value{
A function with rate limiting added
}
\description{
Takes any other function in the Companies House documentation and returns the same function with rate limiting.
Uses \code{trace} to listen in on the headers of each HTTP request. As it's quite noisy, consider running with
\code{suppressMessages()}.
}
\examples{
\donttest{
PersonSearch.Rated <- RateLimitWrap(PersonSearch_limit_first)

PersonSearch.Rated(person, mkey)
PersonSearch.Rated(person, mkey, silent = TRUE)
suppressMessages(PersonSearch.Rated(person, mkey))
}
}
