% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wi_igraph2tnet_function.R
\name{wi_igraph2tnet}
\alias{wi_igraph2tnet}
\title{Weighted indegree igraph2tnet}
\usage{
wi_igraph2tnet(gs, alpha)
}
\arguments{
\item{gs}{igraph object}

\item{alpha}{alpha value}
}
\value{
Dataframe of weighted outdegree centrality results
}
\description{
This function calculates tnet weighted indegree centrality directly on an igraph object and return a dataframe
}
