% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wo_igraph2tnet_function.R
\name{wo_igraph2tnet}
\alias{wo_igraph2tnet}
\title{Weighted outdegree igraph2tnet}
\usage{
wo_igraph2tnet(gs, alpha)
}
\arguments{
\item{gs}{igraph object}

\item{alpha}{alpha value}
}
\value{
Dataframe of weighted outdegree centrality results
}
\description{
This function calculates tnet weighted outdegree centrality directly on an igraph object and return a dataframe
}
