% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompanyNetworkPLOT_function.R
\name{CompanyNetworkPLOT}
\alias{CompanyNetworkPLOT}
\title{One-Mode Company Network}
\usage{
CompanyNetworkPLOT(coynoLIST, mkey, start, end, LABEL, NodeSize = 6)
}
\arguments{
\item{coynoLIST}{list of company numbers}

\item{mkey}{Authorisation key}

\item{start}{Start Year}

\item{end}{End Year}

\item{LABEL}{Node Label - TRUE/FALSE}

\item{NodeSize}{Node Size - default is 6, state number or CENTRALITY (weighted centrality)}
}
\value{
One-Mode Company Network - igraph object
}
\description{
This function creates an company to company network from a list of company numbers, the one-mode projection of the interlocking directorates network
}
