% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectorNetworkPLOT_function.R
\name{DirectorNetworkPLOT}
\alias{DirectorNetworkPLOT}
\title{One-Mode Directors Network}
\usage{
DirectorNetworkPLOT(coynoLIST, mkey, LABEL, NodeSize = 6, start, end)
}
\arguments{
\item{coynoLIST}{list of company numbers}

\item{mkey}{Authorisation key}

\item{LABEL}{Node Label - TRUE/FALSE}

\item{NodeSize}{Node Size - default is 6, place number or CENTRALITY (weighted centrality)}

\item{start}{Start Year}

\item{end}{End Year}
}
\value{
One-Mode Director Network - igraph object
}
\description{
This function creates an director to director network from a list of company numbers, the one-mode projection of the interlocking directorates network
}
