% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterlockNetwork_function.R
\name{InterlockNetwork}
\alias{InterlockNetwork}
\title{Interlocking Directorates Network}
\usage{
InterlockNetwork(coynoLIST, mkey, start, end)
}
\arguments{
\item{coynoLIST}{list of company numbers}

\item{mkey}{Authorisation key}

\item{start}{Start Year}

\item{end}{End Year}
}
\value{
Two-Mode Interlocking Directorates Network - igraph object
}
\description{
This function creates an interlocking directorates network from a list of company number
}
