% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeInterlock_function.R
\name{make_interlock}
\alias{make_interlock}
\title{Make an interlocking directorates network}
\usage{
make_interlock(DF)
}
\arguments{
\item{DF}{Dataframe - column 1 Companies, Column 2 - Directors}
}
\value{
Two-Mode Interlocking Directorates Network - igraph object
}
\description{
This function creates an interlocking directorates network from a dataframe with company and director data
}
